Create unit tag with first related test
Export function TestProviderVersion from api_test to provider_unit_test
Expand help message in api_test to include unit and main "make" commands
